feat: implementa teste de intregação 'deveCadastrarAeroporto'

+++ app/src/main/java/airport/dto/DadosAtualizacaoAeroporto.java
@@ -9,8 +9,8 @@ public record DadosAtualizacaoAeroporto(
         String codigoIata,
         String cidade,
         String codigoPaisIso,
-        BigDecimal latitude,
-        BigDecimal longitude,
-        BigDecimal altitude
+        Double latitude,
+        Double longitude,
+        Double altitude
 ) {
 }
